% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfPetNameCounts}
\alias{dfPetNameCounts}
\title{Pet name counts}
\format{
A data frame with 20790 rows and 3 columns.
\describe{
  \item{Species}{The species of the pet animal.}
  \item{Name}{Name of pet animal.}
  \item{Count}{Number of times the pet name occurrs in the data source.}
}
}
\source{
\url{https://catalog.data.gov/dataset/seattle-pet-licenses}
}
\usage{
dfPetNameCounts
}
\description{
A data frame of animal pet names and corresponding counts of occurrence in the data source.
The following species are present: "Cat", "Dog", "Goat", "Pig".
}
\details{
The pet names are taken from publicly available data of pets license registrations
in the years 2015-2020 in Seattle, WA, USA.
For extensive data analysis see: Anton Antonov, "Pets licensing data analysis", (2020), WordPress.
}
\keyword{datasets}
