% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_binary_data.R
\name{generate_binary_data}
\alias{generate_binary_data}
\title{Generate Binary Data}
\usage{
generate_binary_data(n, prob = 0.5)
}
\arguments{
\item{n}{The number of data points to generate.}

\item{prob}{The probability of success (1). Default is 0.5.}
}
\value{
A numeric vector of 0s and 1s representing binary data.
}
\description{
This function generates random binary data with a specified probability of success.
}
\examples{
generate_binary_data(100, prob = 0.7)
}
