% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_custom_distribution_data.R
\name{generate_custom_distribution_data}
\alias{generate_custom_distribution_data}
\title{Generate Custom Distribution Data}
\usage{
generate_custom_distribution_data(n, pdf, lower, upper)
}
\arguments{
\item{n}{The number of data points to generate.}

\item{pdf}{A function representing the probability density function of the custom distribution.}

\item{lower}{The lower bound of the distribution.}

\item{upper}{The upper bound of the distribution.}
}
\value{
A numeric vector representing data from the custom distribution.
}
\description{
This function generates random data from a custom distribution defined by a probability density function.
}
\examples{
custom_pdf <- function(x) { ifelse(x >= 0 & x <= 1, 2 * x, 0) }
generate_custom_distribution_data(100, custom_pdf, 0, 1)
}
