% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_numeric_data.R
\name{generate_numeric_data}
\alias{generate_numeric_data}
\title{Generate Numeric Data}
\usage{
generate_numeric_data(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{The number of data points to generate. Must be a positive integer.}

\item{mean}{The mean of the generated data. Default is 0.}

\item{sd}{The standard deviation of the generated data. Must be non-negative. Default is 1.}
}
\value{
A numeric vector with the generated data.
}
\description{
This function generates random numeric data with a specified mean and standard deviation.
}
\examples{
# Generate 1000 data points with a mean of 5 and a standard deviation of 2
generate_numeric_data(1000, mean = 5, sd = 2)
# Generate 1000 data points with a mean of 10 and a standard deviation of 3
generate_numeric_data(1000, mean = 10, sd = 3)
}
