% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_poisson_data.R
\name{generate_poisson_data}
\alias{generate_poisson_data}
\title{Generate Poisson Data}
\usage{
generate_poisson_data(n, lambda = 1)
}
\arguments{
\item{n}{The number of data points to generate. Must be a positive integer.}

\item{lambda}{The rate parameter (lambda) of the Poisson distribution. Must be non-negative. Default is 1.}
}
\value{
A numeric vector representing Poisson-distributed data.
}
\description{
This function generates random data following a Poisson distribution with a specified lambda.
}
\examples{
generate_poisson_data(100, lambda = 3)
}
