% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_time_series_data.R
\name{generate_time_series_data}
\alias{generate_time_series_data}
\title{Generate Time Series Data}
\usage{
generate_time_series_data(n, trend = 0, noise_sd = 1)
}
\arguments{
\item{n}{The number of data points to generate. Must be a positive integer.}

\item{trend}{The linear trend component to be added to the time series. Default is 0.}

\item{noise_sd}{The standard deviation of the noise component. Must be non-negative. Default is 1.}
}
\value{
A numeric vector representing the generated time series data.
}
\description{
This function generates random time series data with a specified linear trend and noise level.
}
\examples{
# Generate 100 data points with a trend of 0.5 and noise standard deviation of 1
generate_time_series_data(100, trend = 0.5, noise_sd = 1)
# Generate 100 data points with a trend of 1 and noise standard deviation of 2
generate_time_series_data(100, trend = 1, noise_sd = 2)
}
