% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomDataFrameGenerator.R
\name{RandomDataFrame}
\alias{RandomDataFrame}
\title{Random data frame}
\usage{
RandomDataFrame(
  nrow = NULL,
  ncol = NULL,
  columnNames = NULL,
  columnNamesGenerator = NULL,
  form = "wide",
  generators = NULL,
  minNumberOfValues = NULL,
  maxNumberOfValues = NULL,
  rowNamesQ = FALSE
)
}
\arguments{
\item{nrow}{Number of rows.
If NULL a \code{rpois(1,20)} is used.}

\item{ncol}{Number of columns.
If NULL a \code{rpois(1,7)} is used.
Ignored if \code{columnNames} is not NULL.}

\item{columnNames}{Column names.
If NULL random strings are used.}

\item{columnNamesGenerator}{Generator of the column names.}

\item{form}{The form of the generated data frame, one of "Long", "Wide", or NULL.
If NULL the form is randomly chosen.}

\item{generators}{Generator specifications for the column values.}

\item{minNumberOfValues}{Minimum number of non-NA values.}

\item{maxNumberOfValues}{Maximum number of non-NA values.}

\item{rowNamesQ}{Should the generated data frame have row names or not?}
}
\description{
Generates a random data frame using different specifications.
}
