% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFunctions.R
\name{RandomDate}
\alias{RandomDate}
\title{Random dates}
\usage{
RandomDate(size = 1, min = NULL, max = NULL, ...)
}
\arguments{
\item{size}{Number of words.}

\item{min}{Minimum date.
If NULL then it same as the \code{origin}.}

\item{max}{Maximum date.
If NULL then it is the same as "2100-01-01".}

\item{...}{Additional arguments for \code{\link{as.POSIXct}}}
}
\description{
Generates a vector of random POSIXct dates.
}
\details{
If numerical values are given to the arguments
\code{min} and \code{max} then an \code{origin} argument has to be supplied.
}
