% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFunctions.R
\name{RandomWord}
\alias{RandomWord}
\title{Random words}
\usage{
RandomWord(size = 1, type = "Any")
}
\arguments{
\item{size}{Number of words.}

\item{type}{Type of words.
One of \code{c("Any", "Known", "Common", "Stopword")}.
If NULL is the same as "Any".}
}
\description{
Generates a vector of random words.
}
