% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformers.R
\name{SplitVariableValues}
\alias{SplitVariableValues}
\title{Split variable values}
\usage{
SplitVariableValues(
  data,
  variableToSplit,
  splitPattern = ";",
  idColumn = "ID",
  variableColumn = "Variable",
  valueColumn = "Value"
)
}
\arguments{
\item{data}{A data frame.}

\item{variableToSplit}{A character vector.}

\item{splitPattern}{Split pattern}

\item{idColumn}{Identifier column.}

\item{variableColumn}{Name of the column that has the variables.}

\item{valueColumn}{Name of the column that has the values.}
}
\value{
A data frame.
}
\description{
For a given data frame splits the values of specified variables.
It is assumed that the given data frame has ID, Variable, and Value columns.
(I.e. the data frame is in long form.)
}
